Reworked the way the shop redraws, added things
+The shop runs parallel to the redraw function.

+When the shop redraws, it can still accept payments!

+Little dots at the top of the screen.
The red dot:
Every 10 seconds will appear to notify the user the shop is updating stocks and redrawing.  It also works good as a sort of heart-beat.
The blue dot:
When the shop receives a payment, the blue dot will pop up.  It will also pop up upon receiving a websocket message when it is redrawing.  If you don't know what a websocket is, don't worry about it, it's not really too important anyways.

+When an item is selected, it stays on the screen for 3 "redraw cycles" and then is deselected. (Used to be reset after every redraw).